move ports logic to api calls
